Fix test failures without '.' present in '@INC'
Since Perl 5.26.0, under default conditions, '.' is no longer in @INC

This issue is however presently hidden, as unless somebody explicitly
sets PERL_UNSAFE_INC=0:
- CPAN tools set PERL_USE_UNSAFE_INC=1 when running "./Makefile.PL"
- Test::Harness sets PERL_USE_UNSAFE_INC=1

Its anticipated that in a future perl version, the PERL_USE_UNSAFE_INC
mechansim will go away, leading to these failures.

Without this patch, the issue is always visible when testing via:

  perl Makefile.PL
  make
  perl -Mblib t/foo.t

But is not visible with a standard

  make test

( A good way to see where it gets turned on automatically is put a
  throw-away test somewhere with the code:

  die "BAD PROGRAMMER, NO COOKIE" if $ENV{PERL_USE_UNSAFE_INC};

Then it barfs every time your toolchain hid this problem from you.

Just don't ship that in a test or you will have a bad day )

Bug: https://github.com/jettero/statistics--basic/pull/3